+++ docs/src/markdown/changelog.md
@@ -5,6 +5,7 @@
 - **NEW**: `version` and `version_info` are now accessible via the more standard form `__version__` and `_version_info__`. The old format, while available, is now deprecated.
 - **FIX**: Fix issue where entities in the form `&#35;` would trigger MagicLink's shorthand for issues.
 - **FIX**: Don't install tests when installing package.
+- **FIX**: Fix for BetterEm case `**Strong*em,strong***`.
 
 ## 6.0.0